Main and Inactive View (#3931)
* Added nPMI Plugin Backend Files

Added files for the nPMI plugin backend. This is not yet completely
working.

* Added Comment to Active Function

Commented the isActive function of the plugin.

* Fix Plugin Test File

The test file did not work correctly as of some indentation issue. THis
has been fixed now.

* Summary Tests

Added tests for the summary writer.

* Corrected Indentation for Main File

Corrected the indentation for the main plugin backend file.

* Added Missing License

Added a missing license disclaimer to the summary test file.

* More Indentation Correction

Corrected more of the indentation issues in the python files.

* Further Linter Issues

Fixed further linter issues.

* Linting Correction

Further Linting Correction

* Further Linting

More linting changes.

* Fixed BUILD File

Fixed linter error for BUILD file.

* Fixed Test Issue on TF 1.15

Fixed failing tests because of wrong imports on TF 1.15.

* Rename NPMI to Npmi in Classes

For consistency, this renaming has been done.

* Missed Rename

Missed renaming in one place.

* More Explanation for the Plugin

Added more explanation for the plugin with a readme and one more comment.

* Removed Keys and added Comments

Removed some keys that are not really needed for the plugin and added
more comments to clarify routes.

* Build Deps Cleaned

Cleaned up a little in the build deps.

* Removed Unused Imports

Had several imports that were not used. Removed them now.

* Added Dependency

Added a missing dependency for the plugin.

* Changed dependencies to actual Import

Changed dependecies of the summary writer to the actual imports of this
file.

* Restructured Summary Writer

Summary writer now with three different functions for the routes.

* Fixed Linter Issues

After this restructure, I introduced some linter issues, which are now
fixed.

* Added Safe JSON Encoder

Added a encoder that safely handles numpy nan, inf, and -inf values.

* Linting Fixes

After adding the SafeEncoder, needed some linter fixes.

* Final Linter Change

One more Linter change for the safe encoder.

* Added Proto for Metadata

Added aproto for the metadata field.

* Metadata is Own File

Now, the metadata is its own file. This had to be done for the metadata
content is a proto now and thus metadata is accessed at different
places.

* Begin data_provider Transition

Beginning the transition from multiplexer to data_provider.

* Extra Whitespace Bug

Removed an extra whitespace where there should not be one.

* Plugin Name from Metadata

Now getting the plugin name from metadata to avoid duplication.

* Fix BUILD File

Some fixes in the BUILD file for the CL.

* Added Version Field

Added a version field to the metadata.

* Switch from Multiplexer to Data Provider

Switched from multiplexter to using data provider for serving data.

* Linting

Corrected linting issues.

* More Linting

More linting corrections.

* Removed Safe Encoder for Parse Function

Now parsing the values that we get from a tensor and converting nan to
None.

* Fixed Linting in Convert Function

The convert function now is correctly formatted as well.

* Minor Comment Changes

Changed some comments for better explanation.

* Removed Table Field

Removed the table field from all the data.

* Linting Fix

Fixed a minor linter error.

* Moved To Correct Directory

The backend has been moved to the tensorboard/plugins directory.

* Build Lint Fix

* Fixed Wrong Import

The default.py was still using the old path.

* Added The bare-bones of the nPMI Frontend

Added just the main comopnent without any data fetching whatsoever.

* Remove Duplicate Plugin Class

Removed a duplicate declaration of the npmi plugin class in plugin.py.

* NPMI -> Npmi

To match the backend way of writing this.

* Started Test Implementation

Started implementing tests for the frontend.

* Added Test to Webapp

* Added Tests for NPMI

Added tests for the npmi component.

* BUILD Lint

* BUILD Lint Fix

* Prettier Formatting

* Fixed Issues and Adressed Stephans Comments

Fixed an issue with the a failing test and adressed Stephans PR
comments.

* Require Runs

Runs now required to be passed.

* Corrected Usage of AppState

Corrected usage of State to match the model for tb plugins.

* Missing Dependency

Added a missing dependency which is required for the run selector.

* BUILD Lint

* Add Inactive and Main View

Added a view for when the plugin is inactive, and one as a main starting
view of the plugin.

* License Header

* HTML and SCSS Formatting

* Added Tests for new Modules

* Removed Unused CSS

* Addressed more PR comments

* Fixed Selector in Test

Fixed an incorrect selector name in a test.

* Renamed Selector for Inactive Component