data-loader-behavior: define `reload` on interface
Summary:
A component that extends `DataLoaderBehavior` gains a `reload` method.
This patch declares that on the interface so that clients can call
`reload` even if the component doesn’t override it. Extracted from #3659
as this is needed by both histograms and distributions, and this change
is safer to land.

wchargin-branch: dlb-reload
wchargin-source: cb9b95212897f0254dc62a668a8693dc1c5e5cdd