refactor: tease out plugin enumeration from plugins_module
plugins_module is responsible for providing plugins registry and view
for rendering a plugin. Now, this module should not have enumeration of
all possible plugins in TensorBoard because whether an application wants
to embed a plugin or not is an application logic. For instance, we do
not intend to show debugger_v2 in TB.dev for now.