ng: dispatch changePlugin action when store change
This change on purpose introduces one of the anti-pattern of ngrx usage
in our app: composite action.

Background: plugins are rendered when user interacts with the plugins
selector on the header and plugins know to fetch data (if empty) when
`changePlugin` action fires. However, there is other condition in
which the activePlugin can change: (1) user launched with specific
plugin id deep link or (2) plugins listing is loaded. In the past, for
plugins to fetch the data on load, it had to write necessary code for
detecting such nuances which has been a source of pain and a source of
bug.

This change: in the core effects, that understands the core reducers the
best, detects the condition in which the initial non-null value of
activePlugin is set to dispatch `changePlugin` action without user
interaction on the header.

Alternative design: we can subscribe to the store value and listen to
the change but it is actually no different than other composite action:
a state only changes when an action fires and listening to a change in
the store essentialy creates a chain between multiple actions (e.g.,
action A -> state change -> action B -> state change -> ...).