polymer3: remove `readOnly` from initialized props (#4041)
Summary:
In Polymer 3, using `@property({readOnly: true})` on a property that
also has a TypeScript initializer is always a silent error: the Polymer
decorator will add a getter, causing the initialization code added to
constructor by `tsc` to be a no-op. This patch removes all such
occurrences, as identified by running

```
git grep readOnly \
    'tensorboard/plugins/*polymer3*.ts' \
    'tensorboard/components_polymer3/*.ts' \
    ;
```

and checking manually for initializers.

Test Plan:
Not manually tested. As for completeness, with `-A4` added to the above
`git grep`, the only `=`-sign is a false positive.

wchargin-branch: polymer3-no-readonly