polymer3: remove test/benchamrk build file
Soon, we will remove `tensorboard/components/` while build targets in
the polymer3 test folders are pointing at packages in the polymer2
directory. Since the tests are disabled via tags = ["manual"] already
and they are only causing trouble at this point, we will remove the
BUILD files.

In the future when we rewrite and restore these tests, we can add the
BUILD file back with correct definition.