docs: publish “Generic data authentication” design doc (#4066)
Summary:
This design doc describes extending the data provider APIs to support
authentication, and integrating this support into the TensorBoard web
stack. It relates to the changes in #3816, #3824, and related PRs.  The
main value is in the “Alternatives considered” section, since many
reasonable-sounding options ended up not being viable.

wchargin-branch: docs-design-generic-data-auth