Include HParams in "Getting started with TensorBoard.dev" colab. (#3906)
TensorBoard.dev has launched support for the HParams Dashboard. We want to include some basic HParams logging in the TensorBoard.dev colab.

Update the "Getting started with TensorBoard.dev" so that it includes basic HParams logging in the call to Model.fit().

The HParams output is not very interesting -- we are still only doing a single run -- but we decided it was more important to keep the colab brief. We provide a link to the HParams Dashboard colab for users who want to explore further.