core: fix `--window_title` in Angular TensorBoard (#4099)
Summary:
This behavior is slightly different from Polymer TensorBoard, in that if
the window title is set but then the server restarts and the window
title is no longer set, it will revert to “TensorBoard” instead of
retaining its previous value.

This implementation uses a new view instead of an effect so that we can
read the environment information from the store without violating our
policies about composite actions.

Test Plan:
Launch TensorBoard with `--window_title one`, and note that the title is
properly set. Keep the frontend open, and relaunch the server with
`--window_title two`. Press the in-app “Reload” button, and verify that
the title updates. Finally, relaunch the server with no `--window_title`
and verify that the frontend resets the title to “TensorBoard”.

No TypeScript tests because I can’t think of anything that would go
wrong without being caught by the type system.

wchargin-branch: ng-window-title