cleanup: break projector dependency on weblas
We were using weblas for the matrix multiplication in GPU. Since the
library is not maintained regularly and depends on vulnerable versions
of libraries, we break the dependency to weblas by replacing it with
TensorFlow.js.

When the demo was run with tf.js, I did not see anything out of
ordinary.