projector: fix UMAP dependency issue (#4157)
While doing Polymer 3 migration, we changed everything to use JavaScript
module system (as opposed to tf_web_library). In doing so, we forgot
about the UMAP dependency but TSC did not complain about it because of
our d.ts file.

This change fixes the import and also remove the d.ts file.