feat: create reusable worker
This is one of the foundational library for the new upcoming feature:
Gpu based line chart. In the new line chart, we would like to leverage
OffscreenCanvas which allows worker thread to manipulate the canvas.

In doing so, we are creating a small library that maintains worker
instances so we don't spawn off unbounded nubmer of workers.