line-chart: create reusable worker (#4231)
This is one of the foundational library for the new upcoming feature:
Gpu based line chart. In the new line chart, we would like to leverage
OffscreenCanvas which allows worker thread to manipulate the canvas.

In doing so, we are creating a small library that maintains worker
instances so we don't spawn off unbounded number of workers.

This reverts commit 5c82a480f63a696d511abccb4deb41163928742c.