cleanup: NO_SCHEMA_ERROR to avoid noise in test (#4235)
Angular test specs have their own way of compiling and specifying
modules. As a result, when building Angular template, it can resolve to
a different implementation or nothing (e.g., <mat-icon> without
importing MatIconModule resolves to nothing). This makes Angular's *Ivy*
compiler write warning (instead of throwing exception and halting the
build) and litters our test.

Ideally, there should be a way to assert on console.warn and fail the
test but we will not invest much time in infrastructure in this PR, at
least.