Support followup actions in alert snackbar (#4257)
The Alert feature includes a snackbar view, which
renders the alert message. This change adds a 'dismiss'
button to make all alerts close-able on user gesture.

Clients of AlertActionModule may now register a followup
action, which can generate a dispatched action (along
with payload from the store) when the user clicks on the
followup action label.