Show alert on fetch run failures (#4266)
The recently introduced Alerts ngrx feature offers an affordance
for modules to register alerts. This registers the `fetchRunsFailed`
action to show a message to the user when runs fail to fetch.

In non-Ivy Angular, the entryComponents is required so that
the component factory will be generated, even though it is loaded
on-demand (by snackbar).