line chart: add coordinator (#4248)
Coordinator is a module that converts coordinate system. It is a
stateful wrapper around scale and abstract out some quirks about
different coordinate systems of different renderers.

One feature it has is, it provides some ability to check whether
converted coordinate has to be updated with the updateIdentifier: it
currently is just the time (in number) of an update. This is used later
when determining the coordinate converted data has to be updated.