bug: prevent double data fetch on stamp (#4269)
background: when launching TensorBoard and when inspecting the network
tab on your favorite dev tools, you will notice multiple calls for runs,
tags, and, maybe, scalars. This happened because our Polymer based
dashboards have logic to reload the data on its `ready` lifecycle. On
top of it, on the Angular side, we have a logic to reload the data when
timestamp for plugins list loaded changes for our automatic reloads.

Roughly speaking, there are two ways to solve above problem: (1) make
Angular not invoke reload on new dashboard creation or (2) disable
reload on ready and make sure Angular invokes the reload on the right
condition. In this change, we decided to take the second approach since
we would like to move logic more to the Angular side and we would like
to make it consistent with other plugin types.

Specifically, in this change, we more carefully imperatively invoke the
`reload` by detecting the right condition.