line chart: introduce renderer (#4249)
Renderer is an abstraction over how we are going to draw line, text, and                                                   
rect. For now, we only support two renderer, SVG and THREE.js based                                                        
WebGL.                                                                                                                     
                                                                                                                           
It is worth to note that both renderers we are supporting uses Object to                                                   
represent a pixel--SVG uses elements to describe path, circle, and text and                                                
THREE.js uses Object3d to represent shader. Contrary to canvas renderer that                                               
renders pixels, for an update, the object based renderer allow us make                                                     
modification without erasing entire scene/canvas/DOM to guarantee correctness.                                             
This allows us to make optimization where we minimally make changes to an                                                  
object and let the library/browser make an efficient update to pixels.                                                     
                                                                                                                           
Based on this idea, our renderer in its shape rendering API takes an object as                                             
a parameter to update an object instead of creating one.