rust: add `crc` dependency (#4281)
Summary:
The `crc` module computes CRC checksums. Potential alternatives include
the `crc32` and `crc32fast` crates. But `crc32` is unmaintained and not
widely used, and `crc32fast` does not have Castagnoli tables. So `crc`
it is: the API is simple and it gets the job done.

Test Plan:
That `bazel build //third_party/rust:crc` succeeds suffices.

wchargin-branch: rust-dep-crc