rust: implement bipartite reservoir sampling
Summary:
We introduce a reservoir sampling data structure with a “commit”
operation. The reservoir accepts records into its internal state, but
the records are not directly accessible to clients. When the reservoir
is _committed_ to an external vector, the reservoir moves pending events
out of its internal storage, and reconciles with the external vector to
account for records that have been evicted. This separation enables a
file-reading worker to continually offer records to a reservoir over
which it has exclusive ownership, without contention from readers except
at the defined commit points.

The reservoir is designed to support preemption, but, for ease of code
review, the implementation of preemption is deferred to a later patch.

Test Plan:
Unit tests included.

wchargin-branch: rust-bipartite-reservoir
wchargin-source: fafe42ea89ff0f50307d07675fd6f8f63bb8a37e