workspace: upgrade to newer `rules_rust` (#4307)
Summary:
We pull in bazelbuild/rules_rust#468 to make it easier to build protos,
while not pulling in bazelbuild/rules_rust#463, because that would
require us to update Bazel to 3.0.0. Happily, the commit that we need
merged immediately before the commit that we don’t want! (Yes, this is a
wee bit precarious; we’re working on the Bazel upgrade: #4277.)

Test Plan:
With both Bazel 2.1.0 and Bazel 3.7.0, everything builds, all tests
pass, and TensorBoard looks okay enough.

wchargin-branch: rules-rust-20201105a