line chart: data drawable, a class that renders (#4283)
This change introduces a concept of a drawable. A drawable is a region
in a view where an implementation can draw shapes using PaintBrush with
coordinate converted data.

The DataDrawable abstract class manages data (and its coordinate system
conversion) and render caching on behalf of a subclass. As a subclass,
you only need to implement `redraw` and focus on drawing operations
taking coordinate system converted data and render cache maintenance.