rust: add documentation server (#4302)
Summary:
The Rust Bazel rules provide a `rust_doc` rule for documentation, but it
generates a zipfile with long file paths, which is inconvenient to
consume. This patch adds a simple Werkzeug server to serve from the
generated documentation. It’s compatible with `ibazel run`: just change
a Rust source file and refresh your browser tab to see your changes.

I have a [pending upstream patch for this][upstream], but I’m still
working on it, and it’s on my back burner. In the meantime, this is
helpful for both Rust authors and Rust reviewers on TensorBoard.

[upstream]: https://github.com/bazelbuild/rules_rust/pull/475

Test Plan:
Run `bazel run //tensorboard/data/server:rustboard_core_doc_server` and
navigate to port 6005:

![Screenshot of the docs for `rustboard_core::masked_crc::MaskedCrc` in
a standard rustdoc output format][ss]

[ss]: https://user-images.githubusercontent.com/4317806/98600750-e788bd80-2292-11eb-9f09-fd51ed5e5996.png

wchargin-branch: rust-doc-server