rust: compile protocol buffer bindings with Prost (#4311)
Summary:
This patch adds infrastructure to compile Rust bindings for our protocol
buffer types. The compilation occurs in the build graph via a genrule,
and then the generated files are reflected into the source tree. A Bazel
test verifies that they are kept in sync.

We reflect the generated files into the source tree so that IDE tooling
like `rust-analyzer` can analyze these files more correctly and quickly.
There are ways to get `rust-analyzer` to consider generated files, but
they’re primarily meant for integration with Cargo `build.rs`, not Bazel
(though [Bazel `rules_rust` support for `rust_analyzer` is on the
roadmap][rr-ra]). Such configurations also force `rust-analyzer` to
execute the genrule-equivalent, which isn’t instant, especially when the
build dependencies have not yet been built.

Also, `rules_rust` doesn’t yet seem to support using generated files in
a `rust_library`/`rust_binary`, so this is necessary for now, anyway.

We compile the bindings in the build graph, rather than just providing a
binary that runs the compilation process, primarily because this gives
better control over the execution environment. The data dependencies
here are a bit finicky, and the genrule environment is well defined. By
contrast, a binary could be run either via `bazel run` or `bazel-bin/`,
and either at the repo root or the package directory. By using a
genrule, we only have to support one configuration.

Using a genrule also wins slightly better caching behavior: a normal
build of our Rust library doesn’t actually need to pull in `prost-build`
(though running the tests does).

The generated files are not subject to `rustfmt` because of how they’re
`include!`d into the crate. This is kind of a happy accident. If ever
`rustfmt` learns to format these files, we can fend it off again by
adding `#[rustfmt::skip]` to the module.

[rr-ra]: https://github.com/bazelbuild/rules_rust/issues/378

Test Plan:
Run `//tensorboard/data/server:update_protos_test` and note that it
passes. Then, modify `tensorboard.pb.rs` and re-run it. The test should
fail and instruct you to run `//tensorboard/data/server:update_protos`
to correct the discrepancy. Do so; the test should pass again.

Also, note in the documentation server that the `proto::tensorboard`
module exists and is well populated.

wchargin-branch: rust-compile-protos