rust: add TFRecord reading module (#4306)
Summary:
We implement a module for reading a file that contains a growing stream
of TFRecords. The record contents are opaque; we don’t require that they
be protos or anything else. The module can detect a partially written
record and wait until more data is written to the file.

Test Plan:
Unit tests included.

wchargin-branch: rust-tf-record