refactor: runs-table from mat-table to ngFor (#4286)
This change is in preparation for potential virtual scrolling in the
runs table.

About virtual scroll: Run selector can possibly render many runs; in
real user cases, the run selector had to render more than a thousand
runs. While minor in the grand scheme of things, Angular rendering
thousands of elements with nested components can contribute to AFT or
initial meaning paint.

Virtual scroll module does not seemingly support mat-table and it
requires the items to be rendered via cdkVirtualFor. This change, for
now, renders using ngFor which is API compatible with cdkVirtualFor.