rust: add event file reading module (#4315)
Summary:
We implement a module for reading a TensorFlow event file. This wraps
the TFRecord reading module added in #4307 and parses records as protos.
The reader also stores the last-read event wall time, so that we can
stop reading from event files after a while (cf. `--reload_multifile`).

Along the way, we add a couple utility methods to `TfRecord` that make
it easier to use records in test code.

Test Plan:
Unit tests included.

wchargin-branch: rust-event-file-reading