data: add `DataProvider` gRPC service spec (#4314)
Summary:
This patch adds a gRPC interface that mirrors the `DataProvider` Python
interface. The intent is to define a `GrpcDataProvider` Python type that
is backed by a gRPC service.

The particular service definition here is what I’ve used in drafts of
data provider implementations [in Go][go] and [in Rust][rust], so I can
attest that it’s at least mostly reasonable on both the client and the
server. It’s mostly a one-to-one port of the Python interface and admits
an easy Python client.

[go]: https://github.com/wchargin/tensorboard-data-server
[rust]: https://github.com/wchargin/rustboard-server

Test Plan:
It builds: `bazel build //tensorboard/data/proto:protos_all_py_pb2_grpc`.

wchargin-branch: data-provider-proto