line chart: OffscreenCanvas support for Chart (#4287)
OffscreenCanvas conveniently lets us transfer ownership of Canvas DOM
object to a WebWorker thread so we can reduce expensive point operations
to another process to unlock the UI thread. However, when using
OffscreenCanvas, we have to communicate with the main thread using
message asynchronously.

Since it is not possible to use OffscreenCanvas in some browsers, we
need to gracefully fallback to rendering SVG or WebGL on main thread. As
such, we created an abstraction and use `Chart` and let the Angular side
entry point use the right class.