rust: support reservoir preemption (#4299)
Summary:
This patch extends the bipartite reservoir sampling data structure with
support for preemption. A *preemption* occurs when an record is enqueued
with a step that is not higher than the last-seen step. In a preemption,
all records whose step is not lower than the new record’s step are
evicted.

Test Plan:
Unit tests expanded.

wchargin-branch: rust-reservoir-preemption