line chart: interactive subview (#4291)
This change introduces interactivity to the line chart. The interactive subview
or layer handles mouse and user actions onto the chart and converts it to
mutations.

Understandably, the change is sizable due to all the interactions we are supporting.
If requested, I can break it down so testing refactor and utils are introduced separately.