line chart: export types for consumer of line chart
lib/public_types.ts is not really meant to be imported by consumer of
the line chart as (1) peeking into `lib` folder is a bit messy and (2)
it exports more types than what consumers really need.

The new `types` module is provided as part of the line_chart_v2
dependency and it only contains the types that consumers would need.