rust: create multi-file run loader (#4343)
Summary:
This patch adds a `RunLoader` that reads from multiple event files
within a single run. For now, it only keeps track of scalars, and it
doesn’t do anything with the data that it reads. But it persists
everything that we’ll need for the rest of the pipeline.

Test Plan:
Unit tests included: they create a test log directory and write some
event files, then ensure that they can be read back properly.

wchargin-branch: rust-run-loader