rust: promote event values to `data_compat` (#4349)
Summary:
This patch refactors `run`’s `StagePayload` type into an `EventValue`
type in `data_compat`. This way, we can define enriching conversions
from `EventValue`s to committed values in a layer that’s not specific to
the run loader.

We also implement `initial_metadata` for the `GraphDef` variant, for
symmetry and to make the abstraction clearer. The run loader still
doesn’t handle graphs.

Test Plan:
Unit tests expanded for the new code, and continue to cover the old.

wchargin-branch: rust-promote-eventvalue