feat: reverse fill color for histograms (#4374)
Histogram reverse sorts by step number and, previously, the first step
was the darkest while the last step was the faintest. While this works
for most colors, it does not work well for gray based colors.

Now, it is true that we have a different set of color palettes in time
series and it is less of the problem but now that we let users tweak the
run color, it is easy to imagine user picked color being a problem in
the histograms.

To alleviate the problem, we gave the most important step, the last one,
the most salient color.