rust: commit changes after loading runs (#4352)
Summary:
This patch integrates `RunLoader` and `Commit`. A run loader now commits
all changes at the end of its load cycle. This isn’t sufficient—we also
need it to periodically commit while it’s still loading. But it’s a good
start for now.

Much of the weight is actually in `data_compat`, which owns the logic
for “enriching” raw events into their class-specific representations.
This way, that logic can be reused: e.g., for an alternate consumer that
exposes a stream of all events, without downsampling.

Test Plan:
The new enrichment functionality has dedicated tests, and existing tests
for the run loader have been updated to operate on a `Commit` rather
than poking into the internals.

wchargin-branch: rust-commit-run-loader