webapp: add noscript message (#4404)
Summary:
Since TensorBoard is rendered entirely client-side with Angular and
Polymer, the screen is entirely blank when JavaScript is disabled. This
patch adds a simple “please enable JavaScript” `<noscript>` message.
Fixes #4401.

Test Plan:
In both Chrome and Firefox, the message appears when JavaScript is
disabled but not otherwise. It’s properly centered at wide window sizes
and doesn’t create a horizontal scrollbar at narrow ones. Looks like:

![Screenshot of a blank page with a header “TensorBoard requires
JavaScript”, followed by a message, “Please enable JavaScript and
reload this page.” Both are rendered in Roboto, and appear to be
centered in the screenshot.][ss]

[ss]: https://user-images.githubusercontent.com/4317806/100648627-0b6d7b00-32f6-11eb-9781-68af70d7cb1f.png

wchargin-branch: noscript-message