rust: add nice command-line argument parsing (#4397)
Summary:
This patch outfits `//tensorboard/data/server` with a nice CLI, courtesy
of [`clap`]. The flags are a subset of the main TensorBoard flags, but
with kebab-case rather than snake-case (per everyone-other-than-Google
style, via `clap` defaults). Since we now accept a `--port` argument,
which may be `0`, we also print the port to which the server is bound.

The `clap` approach to argument parsing is “define a struct all of whose
fields implement `FromStr`, then mark it `#[derive(Clap)]` to generate
an argument parser”. This is great on both the static typing and runtime
performance fronts. To add more arguments, we just add more fields to
`struct Opts`. See <https://docs.rs/clap/3.0.0-beta.2> for details (this
syntax is introduced in `clap` v3.x, which is still in beta).

[`clap`]: https://crates.io/crates/clap

Test Plan:
Run `bazel run //tensorboard/data/server` with `-h` and `--help` to see
short and long help text, respectively. Then play around with each
argument in turn.

wchargin-branch: rust-cli