notebook: don’t filter polled instances by PID
Summary:
When the `%tensorboard` cell magic is invoked, we compute a cache key
for the “hermetic environment”, primarily args to `%tensorboard` and the
working directory. We first check whether any running TensorBoard
instances match that cache key, and launch a new instance if none do.
But then, while polling for the new instance to have launched, we had a
different matching criterion, checking for a process ID match instead of
a cache key match.

The idea was that “is this TensorBoard instance’s PID equal to the PID
of the subprocess that we just spawned?” would be a more reliable check.
But on Windows ((╯°□°）╯︵ ┻━┻) this is not the case, presumably because
the `tensorboard` console script has some kind of wrapper process in
certain versions of Python. This manifested as “`%tensorboard` always
times out on the first invocation, but works immediately when I invoke
it again”, since invoking it again triggers the cache key check rather
than the PID check. So we now just check by cache key in all cases, and
the logic is consistent, if a bit less precise overall.

Test Plan:
Still works for me on Linux, with both new and existing TensorBoard
processes across multiple (concurrent) cache keys. Would be great for a
Windows user to confirm repro and fix…

wchargin-branch: notebook-poll-no-pid-filter
wchargin-source: 85ff3ae7dd6ecab3f6580ef0c0611a1e09ef0630