Limit max scalar smoothing to 0.99 / 0.999 (#4372)
The smoothing slider setting affects line charts. In the Scalars
dashboard, the maximum value has traditionally been 0.999, not
1.0. The Time Series dashboard allows the slider to go all the way
to 1, which "flattens the curve" onto the x-axis, often much less
meaningful than a "close to 1" number, which still shows the
general shape of the curves.

This updates the Time Series' scalar cards to be use a max
smoothing of 0.99 when dragging the slider, and 0.999 when
entering a value in the input field.