rust: remove some `..Default::default()`s (#4424)
Summary:
In #4399, we decided to recommend `..Default::default()` on struct
initializers for protos whose source of truth is not in our repo. This
patch removes trivial `..Default::default()`s from struct initializers
for `tensorboard.data` protos, which *do* live in our repo. We still use
it sometimes: e.g., some requests have a `run_tag_filter` field that we
don’t always care to set.

This patch will make it harder to add new fields to these protos, but
makes the code a bit easier to read in the meantime.

Test Plan:
Compilation suffices; unit tests are a nice sanity check.

wchargin-branch: rust-remove-default-frus