rust: support loading data server from Pip package
Summary:
With `--load_fast`, TensorBoard can now find the data server binary by
asking the `tensorboard_data_server` Pip package. That Pip package
bundles a platform-specific binary and provides a `server_binary`
function to locate it.

The Pip package can be easily built with Bazel, but it does not require
building with Bazel: you can build the Rust binary however you like
(e.g., with Cargo) and then build the Pip package with plain Python.
This is so that we can easily build Pip packages for Mac and Windows
without having to get Bazel working on those platforms.

Test Plan:
For the Pip package, an end-to-end test is included, but it’s tagged as
a manual test because it depends on the enclosing Python environment, so
run it explicitly:

```
bazel test //tensorboard/data/server/pip_package:build_pip_package_test
```

For an end-to-end test, follow the new instructions in `DEVELOPMENT.md`.
If you want to be extra sure that it’s really running the new binary,
you can add `std::process::exit(77)` to the top of `main.rs` and see
that TensorBoard fails accordingly.

wchargin-branch: rust-tds-pip-package
wchargin-source: 681eda37cf602c6f67b3332831d38b359ac06ee4