ci: build and store data server Pip package
Summary:
A new CI job in our existing workflow builds the RustBoard binary and
its Pip package, and uploads the Pip package as an artifact on the job.
(The package is not uploaded to PyPI or anything like that.)

For now, we only build on Linux. Once the package build script supports
macOS, we should hopefully be able to just add that to the platform
matrix in this job config. Windows will probably need some adjustments
to deal with the shell in the action workflow and the `.exe` suffix on
the binary. We’ll handle that later.

Test Plan:
Check that CI properly builds a package. Download and install it into
your virtualenv. Then, run TensorBoard (with #4429) and check that it
locates the RustBoard server.

wchargin-branch: rust-tds-pip-ci
wchargin-source: 62178846ddb2694fb4f9ccf20ae4a214c7f4a74c